mdrun_multi and grompp_multi have common iteration semantics.
Updated inputs/outputs for grompp_multi and mdrun_multi, so that
outputs are mapped onto individual arrays (consistent with inputs)
and the instance numbers are based on the number of elements in
the input arrays: all arrays that have N>1 elements are used as
iterators (i.e. their elements are distributed over the instances).
These elements must all have the same number of elements.

For now, this is the most straight-forward solution that is
compatible with the concept of a dataflow network (i.e. the state
of the function is fully determined by its inputs).
In the future, we might want to think about treating arrays differently:
perhaps we should do implicit multiple instancing after all.