Heartbeat monitoring moved to originating server
The heartbeat monitor list and the running command list are now the same,
and reside on the command's originating server.
This fixes issues with jobs that are spuriously though of as running by
the originating server: jobs are always either in the queue or running
with a ticking heartbeat monitor.

The heartbeat time is dictated by the worker server, but the checking interval
is set by the originating server.

TODO: - Heartbeat signal relays from worker server to originating server.
      - Persistence.
      - Cleanup and sending back of data at worker server.