Heartbeat requests are forwarded.
Heartbeat requests are now forwarded to the originating server (the
one that owns the command).

This commit fixes a few issues with server to server communication. There
are now two classes in cpc/server/message/server_command for normal server
to server communication and communication to a server that is not (yet) in the
server list.