'msg' value in active instances, with errors, warnings, logs
All active instances now contain a 'msg' subvalue (in addition to
in, out, sub_in, sub_out & the active instances contained in the
instance) that has fields
- 'error' for an error message
- 'warning' for a warning
- 'log' for the log of the instance

This makes it possible to explicitly reference error messages and
to get them through cpcc get, in line with the 'everything-is-a-value'
philosophy.

It also makes errors&warnings part of the dataflow network (although
run-on-update on them is not implemented yet).

We should still implement a history for error messages and warnings. The
current 'error' and 'warning' fields only pertain to the current state
of the active instance.

Change-Id: I0aba25f6729d78ac07df047154aaa92f04c81a5a