Added server configurable env. variable CPC_NUM_THREADS
A new server-configurable environment variable CPC_NUM_THREADS
takes the place of OMP_NUM_THREADS set directly by the server.
This moves the work of limiting core number usage to the functions,
where it must be limited actively. This is currently only relevant
for msm, where it is implemented.

The server config value is: server_cores

Change-Id: I2be2fd2bc9258f8871d57f7d6a3d65ed3cfe3334