Partially fixed issue with server shutdown
The stop-server command exits the thread which handles the request. When
the unverified handler ran in a daemonized thread, only that thread was
shut down which didn't propagate to shutdown the whole process. This
changes so the verified runs in the daemonized thread and the unverified
in the main thread, making the shutdown work properly when issued via
the unverfied channel (i.e. login). If, however, a stop-server is issued
via the verified channel (using bundle), this might still be an issue.

Change-Id: Ie51139d77aac15f4af73a1f54fa8f37e28c87a87