Added a client configuration
clientconfig.cfg stored in .copernicus in now generated when the client
is used. This is (for now) used as a library for server bookmarks. This
means the client can have many servers added and may switch between them
using the use-server command.

The bundle may still be used by supplying it to the client with the -c
flag, allowing for a more secure two-way connection and root access.

Change-Id: If58d5774c78da0a77480feb020582cb5cde2d0a3