Fixed a bug with keep alive
The keep alive used ServerMessage to connec to the target server which
relies on data in the network topology cache. If a server had
estasblished an outgoing connection but not yet updated its topology the
target host would not be found and the keep alive would break.

This uses DirectServerMessage which relies on the data in the node
objects which are kept for all nodes we have a direct link to.

This fixes many failing test cases.

Change-Id: I52c90756413c2d05ba0d1f54eef5c56282d72155